improve and clean up custom slider
