+++ scripts/components/custom-slider/custom-slider-style.js
@@ -1,0 +1,19 @@
+/**
+ * CustomRangeSlider style.
+ *
+ * @param {string} DEFAULT - The default slider.
+ */
+ export const CustomRangeSliderStyle = {
+	DEFAULT: 'default',
+	COLUMN_PICKER: 'columnPicker',
+	COLUMN_PICKER_WITH_DOTS: 'columnPickerWithDots',
+};
+
+/**
+ * CustomSlider style.
+ *
+ * @param {string} DEFAULT - The default slider.
+ */
+ export const CustomSliderStyle = {
+	DEFAULT: 'default',
+};
